Simplify default session template bootstrap
